Fix warnings

+++ TestApp/Source/Win32/App.cpp
@@ -81,10 +81,10 @@ namespace
         {
             return;
         }
-        auto width = rect.right - rect.left;
-        auto height = rect.bottom - rect.top;
+        auto width = static_cast<float>(rect.right - rect.left);
+        auto height = static_cast<float>(rect.bottom - rect.top);
         runtime = std::make_unique<Babylon::RuntimeWin32>(hWnd, rootUrl, width, height);
-        
+
         // issue a resize here because on some platforms (UWP, WIN32) WM_SIZE is received before the runtime construction
         // So the context is created with the right size but the nativeWindow still has the wrong size
         // depending on how you create your app (runtime created before WM_SIZE is received, this call is not needed)
